01-15
Dubble Buffering

+++ GDIPlus_1/GDIPlus_1/GDIPlus_1.cpp
@@ -20,5 +20,5 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance,
 	{
 		return 0;
 	}
-	return (int)mainWnd.MessageLoop();
+	return mainWnd.MessageLoop();
 }
